Refactored into new flightsystems concept.
This patch converts the project to a workspace with two binaries and
one library. One binary will be for avionics and the other for flight
planning. The library will handle interactions with KRPC/KRPC_Mars
as an intermediate interface that abstracts some of the direct
information requests to KRPC. These could eventually be upstreamed
to KRPC Mars.